requirements.txt
torch
torchaudio
pyyaml
cmap
tensorboard